delete the old deploy_brew rule, and rename deploy_tap to deploy_brew
